backend/requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.5
flask-migrate==4.0.4
flask-jwt-extended==4.5.2
python-dotenv==1.0.0
